html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-optional-003.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: tokenization -- legacy position features `screenx`, `screeny`</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">

<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedPostMessage.js"></script>
<script>
var windowURL = 'resources/message-opener.html';

[ 'screenx=142',
  ' screenx = 142',
  'screenx==142',
  '\nscreenx= 142',
  ',screenx=142,,',
  'SCREENX=142',
  'screenX=142'
].forEach((features, idx, arr) => {
  async_test(t => {
    var prefixedMessage = new PrefixedMessageTest();
    prefixedMessage.onMessage(t.step_func_done((data, e) => {
      e.source.close();
      assert_equals(data.left, 142);
    }));
    var win = window.open(prefixedMessage.url(windowURL), '', features);
  }, `"${features}" (${idx + 1} of ${arr.length}) should set left position of opened window`);
});

[ 'screeny=142',
  ' screeny = 142',
  'screeny==142',
  '\nscreeny= 142',
  ',screeny=142,,',
  'SCREENY=142',
  'screenY=142'
].forEach((features, idx, arr) => {
  async_test(t => {
    var prefixedMessage = new PrefixedMessageTest();
    prefixedMessage.onMessage(t.step_func_done((data, e) => {
      e.source.close();
      assert_equals(data.top, 142);
    }));
    var win = window.open(prefixedMessage.url(windowURL), '', features);
  }, `"${features}" (${idx + 1} of ${arr.length}) should set top position of opened window`);
});

</script>
